move Credentials model to its own package
